TRY_CATCH -> TRY+CATCH+END_CATCH, the manual conversions
This fixes the cases where we're using the exception object outside of
a the CATCH block, and cases where we were using "else" after a CATCH.

I originally planned on doing these changes in the preparatory patch
that normalizes TRY_CATCH (the one just before the wholesale scripted
conversion patch), but then that feels like it'd make the resulting
patch harder to read and not as clear.  It's a lot less work this way
too.  :-) I can still do that if people prefer though.  Otherwise, I
plan to squash this with the previous patch, and push it all together.