TRY_CATCH -> TRY+CATCH+END_CATCH everywhere
This converts all TRY_CATCH callers to TRY + CATCH + END_CATCH.

This was all done with a script, no manual editing involved, but I
audited the patch to make sure things were converted as expected.

This is split from the previous patch only for review and maintanance
convencience.  I plan on merging this with the previous patch before
pushing.