more making TRY/CATCH callers look more like real C++ try/catch blocks
All these were caught by actually making TRY/CATCH use try/catch
behind the scenes, which then resulted in the build failing because
there was code between the try and catch blocks.

As with the prior patch, this could either be done _before_ running
the TRY_CATCH conversion script, or be squashed into it.