Linux native: Use TRAP_BRKPT/TRAP_HWBPT
This patch adjusts the native Linux target backend to tell the core
whether a trap was caused by a breakpoint.

It teaches the target to get that information out of the si_code of
the SIGTRAP siginfo.

As I can't be sure that this mechanism works on all supported
architectures, I left the old code in place, guarded with a macro.
We'll remove the fallback code once we're sure all supported
architectures support this method.

For years I assumed that x86 just didn't include this information in
the siginfo, because when a breakpoint trap is raised, si_code does
not show TRAP_BRKPT, as one sees on other architectures.  Because of
that, I had added an entry to gdb's kernel wishlist at
https://sourceware.org/gdb/wiki/LinuxKernelWishList for TRAP_BRKPT in
2012...

But taking another closer look, the x86 port gets odd SIGTRAP si_code
values, but they _are_ usable as is.

Based on reading the kernel's source and experimenting things, this is
how x86's si_code works:

| what                                     | si_code    |
|------------------------------------------+------------|
| software breakpoints (int3)              | SI_KERNEL  |
| single-steps                             | TRAP_TRACE |
| single-stepping a syscall                | TRAP_BRKPT |
| user sent SIGTRAP                        | 0          |
| exec SIGTRAP (when no PTRACE_EVENT_EXEC) | 0          |
| hardware breakpoints/watchpoints         | TRAP_HWBPT |

It's almost as if TRAP_BRKPT and SI_KERNEL were reversed...

Both user-sent traps and execs get the same si_code, but that's not a
problem given PTRACE_EVENT_EXEC.

If the kernel is ever "fixed" to use the right values, we can always
detect it like we detect support for ptrace features (by having gdb
debug itself a bit at startup, and looking at the si_code of a
breakpoint hit).

As I can't be sure that this mechanism works on all supported
architectures, I left the old code in place, guarded with a macro.  I
propose enabling this by default, and only disabling if problems are
reported.  If we only enable on an arch-by-arch basis, it's almost
certain that even if all architectures can use this, we'll never end
up enabling it on all architectures...  If after a while, no problem
is ever reported, we'll remove the fallback code then.  And if some
architecture does indeed trip on some problem, we'll disable the new
code on that arch, but I'll make some loud noises about fixing it in
the kernel too, so later GDB releases can again consider dropping the
old code.

Tested on x86-64 Fedora 20, s390 Fedora 20, and ppc64 Fedora 20.
An ealier version was tested on ARM Fedora 21.


gdb/ChangeLog:
2015-02-25  Pedro Alves  <palves@redhat.com>

	* linux-nat.c (save_sigtrap): Check for breakpoints before
	checking watchpoints.
	(status_callback) [USE_SIGTRAP_SIGINFO]: Don't check whether a
	breakpoint is inserted if relying on SIGTRAP's siginfo.si_code.
	(check_stopped_by_breakpoint) [USE_SIGTRAP_SIGINFO]: Decide whether
	a breakpoint triggered based on the SIGTRAP's siginfo.si_code.
	(linux_nat_stopped_by_sw_breakpoint)
	(linux_nat_supports_stopped_by_sw_breakpoint)
	(linux_nat_stopped_by_hw_breakpoint)
	(linux_nat_supports_stopped_by_hw_breakpoint): New functions.
	(linux_nat_wait_1): Don't re-increment the PC if relying on
	SIGTRAP's siginfo->si_code.
	(linux_nat_add_target): Install new target methods.
	* linux-thread-db.c (check_event): Don't account for breakpoint PC
	offset if the target already adjusted the PC.
	* nat/linux-ptrace.h (USE_SIGTRAP_SIGINFO): New.
	(GDB_ARCH_TRAP_BRKPT): New.
	(TRAP_HWBKPT): Define if not already defined.